Move FreeCommands and SkipCommand to their own file
